Install Spree

+++ db/migrate/20140901220540_add_price_to_releases.rb
@@ -1,5 +1,5 @@
 class AddPriceToReleases < ActiveRecord::Migration
   def change
-    add_column :releases, :price, :double
+    add_column :releases, :price, :money
   end
 end
